+++ docs/sidebars.js
@@ -31,7 +31,8 @@ module.exports = {
       label: 'Data Preparation',
       items: [
         'data-file-types',
-        'data-hosting'
+        'data-hosting',
+        'data-troubleshooting',
       ],
     },
     {
@@ -47,7 +48,8 @@ module.exports = {
       label: 'Developer Docs',
       items: [
         'dev-overview',
-        'dev-add-component'
+        'dev-add-component',
+        'dev-plugins'
       ],
     },
     {
@@ -66,7 +68,11 @@ module.exports = {
       collapsed: false,
       items: [
         'tutorials',
-        'tutorial-visium'
+        'tutorial-visium',
+        'tutorial-gh-pages',
+        'tutorial-plugin-view-type',
+        'tutorial-plugin-coordination-type',
+        'tutorial-plugin-file-type',
       ],
     },
   ],