+++ src/app/view-config-utils.js
@@ -1,13 +1,19 @@
 /* eslint-disable no-plusplus */
 /* eslint-disable camelcase */
+import difference from 'lodash/difference';
+import cloneDeep from 'lodash/cloneDeep';
 import packageJson from '../../package.json';
 import { getNextScope } from '../utils';
 import {
-  DEFAULT_COORDINATION_VALUES,
-  COMPONENT_COORDINATION_TYPES,
   AUTO_INDEPENDENT_COORDINATION_TYPES,
 } from './state/coordination';
-import { CoordinationType } from './constants';
+import { getViewTypes } from './component-registry';
+import { getFileTypes } from '../loaders/types';
+import {
+  getComponentCoordinationTypes,
+  getDefaultCoordinationValues,
+  getCoordinationTypes,
+} from './plugins';
 import { SCHEMA_HANDLERS } from './view-config-versions';
 
 /**
@@ -35,6 +41,7 @@ export function getExistingScopesForCoordinationType(config, coordinationType) {
  * @returns {object} The new view config.
  */
 function coordinateComponentsTogether(config, coordinationType, scopeValue) {
+  const componentCoordinationTypes = getComponentCoordinationTypes();
   const scopeName = getNextScope(getExistingScopesForCoordinationType(config, coordinationType));
   const newConfig = {
     ...config,
@@ -53,7 +60,7 @@ function coordinateComponentsTogether(config, coordinationType, scopeValue) {
         // Only set the coordination scope if this component uses this coordination type,
         // and the component is missing a coordination scope for this coordination type.
         ...((
-          COMPONENT_COORDINATION_TYPES[component.component].includes(coordinationType)
+          componentCoordinationTypes[component.component].includes(coordinationType)
           && !component.coordinationScopes?.[coordinationType]
         ) ? {
           // Only set the new scope name if the scope name
@@ -76,6 +83,7 @@ function coordinateComponentsTogether(config, coordinationType, scopeValue) {
  * @returns {object} The new view config.
  */
 function coordinateComponentsIndependent(config, coordinationType, scopeValue) {
+  const componentCoordinationTypes = getComponentCoordinationTypes();
   const newConfig = {
     ...config,
     layout: [...config.layout],
@@ -84,7 +92,7 @@ function coordinateComponentsIndependent(config, coordinationType, scopeValue) {
   newConfig.layout.forEach((component, i) => {
     // Only set the coordination scope if this component uses this coordination type,
     // and the component is missing a coordination scope for this coordination type.
-    if (COMPONENT_COORDINATION_TYPES[component.component].includes(coordinationType)
+    if (componentCoordinationTypes[component.component].includes(coordinationType)
       && !component.coordinationScopes?.[coordinationType]
     ) {
       const scopeName = getNextScope([
@@ -116,19 +124,23 @@ function initializeAuto(config) {
   let newConfig = config;
   const { layout, datasets } = newConfig;
 
+  const componentCoordinationTypes = getComponentCoordinationTypes();
+  const defaultCoordinationValues = getDefaultCoordinationValues();
+  const coordinationTypes = getCoordinationTypes();
+
   // For each coordination type, check whether it requires initialization.
-  Object.values(CoordinationType).forEach((coordinationType) => {
+  coordinationTypes.forEach((coordinationType) => {
     // A coordination type requires coordination if at least one component is missing
     // a (coordination type, coordination scope) tuple.
     // Components may only use a subset of all coordination types.
     const requiresCoordination = !layout
       .every(c => (
-        !COMPONENT_COORDINATION_TYPES[c.component].includes(coordinationType)
+        (!componentCoordinationTypes[c.component].includes(coordinationType))
                 || c.coordinationScopes?.[coordinationType]
       ));
     if (requiresCoordination) {
       // Note that the default value may be undefined.
-      let defaultValue = DEFAULT_COORDINATION_VALUES[coordinationType];
+      let defaultValue = defaultCoordinationValues[coordinationType];
       // Check whether this is the special 'dataset' coordination type.
       if (coordinationType === 'dataset' && datasets.length >= 1) {
         // Use the first dataset ID as the default
@@ -151,6 +163,58 @@ function initializeAuto(config) {
   return newConfig;
 }
 
+export function checkTypes(config) {
+  // Add a log message when there are additionalProperties in the coordination space that
+  // do not appear in the view config JSON schema,
+  // with a note that this indicates either a mistake or custom coordination type usage.
+  const coordinationTypesInConfig = Object.keys(config.coordinationSpace || {});
+  const allCoordinationTypes = getCoordinationTypes();
+  const unknownCoordinationTypes = difference(coordinationTypesInConfig, allCoordinationTypes);
+  if (unknownCoordinationTypes.length > 0) {
+    return [false, `The following coordination types are not recognized: [${unknownCoordinationTypes}].\nIf these are plugin coordination types, ensure that they have been properly registered.`];
+  }
+  // Add a log message when there are views in the layout that are neither
+  // core views nor registered plugin views.
+  const viewTypesInConfig = config.layout.map(c => c.component);
+  const allViewTypes = getViewTypes();
+  const unknownViewTypes = difference(viewTypesInConfig, allViewTypes);
+  if (unknownViewTypes.length > 0) {
+    return [false, `The following view types are not recognized: [${unknownViewTypes}].\nIf these are plugin view types, ensure that they have been properly registered.`];
+  }
+  // Add a log message when there are file definitions with neither
+  // core nor registered plugin file types.
+  const fileTypesInConfig = config.datasets.flatMap(d => d.files.map(f => f.fileType));
+  const allFileTypes = getFileTypes();
+  const unknownFileTypes = difference(fileTypesInConfig, allFileTypes);
+  if (unknownFileTypes.length > 0) {
+    return [false, `The following file types are not recognized: [${unknownFileTypes}].\nIf these are plugin file types, ensure that they have been properly registered.`];
+  }
+  return [true, 'All view types, coordination types, and file types that appear in the view config are recognized.'];
+}
+
+/**
+ * Assign unique ids for view definitions where
+ * they are missing a value for the uid property
+ * in layout[].uid.
+ * @param {object} config The view config
+ * @returns The updated view config.
+ */
+function assignViewUids(config) {
+  const { layout } = config;
+  const usedIds = layout.map(view => view.uid);
+  layout.forEach((view, i) => {
+    // Assign uids for views where they are not present.
+    if (!view.uid) {
+      const nextUid = getNextScope(usedIds);
+      layout[i].uid = nextUid;
+      usedIds.push(nextUid);
+    }
+  });
+  return {
+    ...config,
+    layout,
+  };
+}
 
 /**
  * Initialize the view config:
@@ -160,12 +224,14 @@ function initializeAuto(config) {
  * Should be "stable": if run on the same view config twice, the return value the second
  * time should be identical to the return value the first time.
  * @param {object} config The view config prop.
+ * @returns The initialized view config.
  */
 export function initialize(config) {
-  if (config.initStrategy === 'auto') {
-    return initializeAuto(config);
+  let newConfig = cloneDeep(config);
+  if (newConfig.initStrategy === 'auto') {
+    newConfig = initializeAuto(config);
   }
-  return config;
+  return assignViewUids(newConfig);
 }
 
 export function upgradeAndValidate(oldConfig) {