+++ src/components/data-hooks.js
@@ -5,13 +5,14 @@ import { useSetWarning } from '../app/state/hooks';
 import {
   AbstractLoaderError,
   LoaderNotFoundError,
+  DatasetNotFoundError,
 } from '../loaders/errors/index';
 import {
   DEFAULT_MOLECULES_LAYER,
   DEFAULT_CELLS_LAYER,
   DEFAULT_NEIGHBORHOODS_LAYER,
 } from './spatial/constants';
-import { DEFAULT_COORDINATION_VALUES } from '../app/state/coordination';
+import { getDefaultCoordinationValues } from '../app/plugins';
 
 /**
  * Warn via publishing to the console
@@ -41,11 +42,12 @@ function initCoordinationSpace(values, setters, initialValues) {
   if (!values || !setters) {
     return;
   }
+  const defaultCoordinationValues = getDefaultCoordinationValues();
   Object.entries(values).forEach(([coordinationType, value]) => {
     const setterName = `set${capitalize(coordinationType)}`;
     const setterFunc = setters[setterName];
     const initialValue = initialValues && initialValues[coordinationType];
-    const shouldInit = equal(initialValue, DEFAULT_COORDINATION_VALUES[coordinationType]);
+    const shouldInit = equal(initialValue, defaultCoordinationValues[coordinationType]);
     if (shouldInit && setterFunc) {
       setterFunc(value);
     }
@@ -623,6 +625,11 @@ export function useRasterData(
 
   useEffect(() => {
     if (!loaders[dataset]) {
+      if (isRequired) {
+        warn(new DatasetNotFoundError(dataset), setWarning);
+      } else {
+        setItemIsReady('raster');
+      }
       return;
     }