+++ src/loaders/anndata-loaders/CellSetsZarrLoader.js
@@ -116,13 +116,13 @@ export default class CellSetsZarrLoader extends AbstractTwoStepLoader {
   loadCellSetIds() {
     const { options } = this;
     const cellSetZarrLocation = options.map(({ setName }) => setName);
-    return this.dataSource.loadObsVariables(cellSetZarrLocation);
+    return this.dataSource.loadObsColumns(cellSetZarrLocation);
   }
 
   loadCellSetScores() {
     const { options } = this;
     const cellSetScoreZarrLocation = options.map(option => option.scoreName || undefined);
-    return this.dataSource.loadObsVariables(cellSetScoreZarrLocation);
+    return this.dataSource.loadObsColumns(cellSetScoreZarrLocation);
   }
 
   async load() {