+++ src/loaders/index.js
@@ -3,3 +3,5 @@ export { default as MatrixZarrLoader } from './MatrixZarrLoader';
 export { default as GenesJsonAsMatrixZarrLoader } from './GenesJsonAsMatrixZarrLoader';
 export { default as ClustersJsonAsMatrixZarrLoader } from './ClustersJsonAsMatrixZarrLoader';
 export { default as AnnDataLoaders } from './anndata-loaders';
+export { default as LoaderResult } from './LoaderResult';
+export { default as AbstractTwoStepLoader } from './AbstractTwoStepLoader';