+++ src/loaders/types.js
@@ -1,4 +1,5 @@
 import { FileType } from '../app/constants';
+import { getPluginFileType, getPluginFileTypes } from '../app/plugins';
 
 import JsonLoader from './JsonLoader';
 import MatrixZarrLoader from './MatrixZarrLoader';
@@ -28,5 +29,20 @@ export const fileTypeToLoaderAndSource = {
 };
 
 export function getSourceAndLoaderFromFileType(type) {
-  return fileTypeToLoaderAndSource[type] || [JsonSource, JsonLoader];
+  if (fileTypeToLoaderAndSource[type]) {
+    return fileTypeToLoaderAndSource[type];
+  }
+  const pluginFileType = getPluginFileType(type);
+  if (pluginFileType) {
+    return pluginFileType;
+  }
+  // Fallback to JSON.
+  return [JsonSource, JsonLoader];
+}
+
+export function getFileTypes() {
+  return [
+    ...Object.keys(fileTypeToLoaderAndSource),
+    ...getPluginFileTypes(),
+  ];
 }